Move projects service into services folder
